requirements.txt
numpy==1.23.5
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
tensorflow-cpu==2.18.0